# backend/models/model_user_login_req.go
/*
 * Assignment2
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package models

type UserLoginReq struct {

	Username string `json:"username"`

	Password string `json:"password"`
}